Ajustes en vistas y correción merge
- Se cambio la posición del botón de retorno en vista SYD
- Se solucionaron inconvenientes merge
