+++ ProyectoINE/Controllers/HomeController.cs
@@ -69,10 +69,28 @@ namespace ProyectoINE.Controllers
                     modeloResultado = decreciente.Calcular(model);
                     vistaDestino = "SaldoDecrecienteTable";
                     break;
-                /*case 4:
+                case 4:
                     // Método de suma de los dígitos de los años
+                    var datosSYD = new DatosSYD
+                    {
+                        B = (double)model.CostoInicial,
+                        VR = (double)model.ValorResidual,
+                        N = model.VidaUtilAnios,
+                        k = model.TasaUso
+                    };
+
+                    var resultado = SYDService.Calcular(datosSYD);
+
+                    if (!string.IsNullOrEmpty(resultado.Error))
+                    {
+                        ModelState.AddModelError("", resultado.Error);
+                        model.Items = ObtenerTiposDeDepr();
+                        return View("Index", model);
+                    }
+                    vistaDestino = "SYDTable";
+                    modeloResultado = resultado;
                     break;
-                case 5:
+                /*case 5:
                     // Método de unidades de producción
                     break;
                 case 6:
