--- ProyectoINE/Models/DatosLineaRecta.cs
@@ -1,9 +1,0 @@
-   {
-  public class DatosLineaRecta
-    {
-        public double B { get; set; }    // Costo inicial
-        public double VR { get; set; }   // Valor de rescate
-        public int N { get; set; }       // Vida útil
-        public int k { get; set; }       // Año específico
-    }
-}
