--- ProyectoINE/Models/DatosSYD.cs
@@ -1,7 +1,0 @@
-public class DatosSYD
-{
-    public double B { get; set; }
-    public double VR { get; set; }
-    public int N { get; set; }
-    public int k { get; set; }
-}
