+++ ProyectoINE/Models/ViewModels/DatosSYD.cs
@@ -1,0 +1,10 @@
+namespace ProyectoINE.Models.ViewModels
+{
+    public class DatosSYD
+    {
+        public double B { get; set; }
+        public double VR { get; set; }
+        public int N { get; set; }
+        public int k { get; set; }
+    }
+}
