+++ ProyectoINE/Views/Home/SYDTable.cshtml
@@ -6,7 +6,9 @@
 
 <div class="container mt-4">
     <h2 class="text-center mb-4">Resultado del Método SYD</h2>
-
+    <br />
+    <a asp-controller="Home" asp-action="Index" class="btn btn-primary">Volver al Inicio</a>
+    <br />
     <table class="table table-bordered table-striped">
         <thead class="table-dark">
             <tr>
@@ -46,7 +48,4 @@
         </tbody>
     </table>
 
-    <div class="text-center">
-        <a asp-action="Index" class="btn btn-primary">Volver al inicio</a>
-    </div>
 </div>
